+++ srcs/CalculateFolderSize.UI.Android/GlobalUsing.cs
@@ -1,0 +1,1 @@
+global using AndroidEnvironment = Android.OS.Environment;