+++ srcs/CalculateFolderSize.UI.Android/MainApplication.cs
@@ -4,6 +4,7 @@ using Avalonia;
 using Avalonia.Android;
 using CalculateFolderSize.Core;
 using CalculateFolderSize.UI.Shared;
+using CalculateFolderSize.UI.Shared.Interfaces;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -13,17 +14,18 @@ using System;
 namespace CalculateFolderSize.UI.Android;
 
 /// <summary>
-/// 应用程序类
+/// 主应用程序类
 /// </summary>
 [Application]
-public class Application : AvaloniaAndroidApplication<App>
+public class MainApplication : AvaloniaAndroidApplication<App>
 {
     /// <summary>
-    /// 初始化一个新的 <see cref="Application"/> 实例
+    /// 初始化一个新的 <see cref="MainApplication"/> 实例
     /// </summary>
     /// <param name="javaReference">Java 参考</param>
     /// <param name="transfer">JNI 处理所有权</param>
-    protected Application(nint javaReference, JniHandleOwnership transfer) : base(javaReference, transfer) { }
+    protected MainApplication(nint javaReference, JniHandleOwnership transfer)
+        : base(javaReference, transfer) { }
 
     /// <summary>
     /// 自定义应用程序构建器
@@ -41,13 +43,17 @@ public class Application : AvaloniaAndroidApplication<App>
             level = LogLevel.Information;
         }
 
-        App.Services = new ServiceCollection()
-            .AddLogging(builder => builder.AddFile(Constants.LatestLogFilePath, o => o.MinLevel = level))
+        var services = new ServiceCollection()
             .AddSingleton<IConfiguration>(configuration)
             .AddCore()
             .AddUIShared()
-            .BuildServiceProvider();
+            .AddSingleton<IStorageAccessService, StorageAccessService>();
 
+        _ = AndroidEnvironment.IsExternalStorageManager
+            ? services.AddLogging(b => b.AddFile(Constants.LatestLogFilePath, o => o.MinLevel = level))
+            : services.AddLogging();
+
+        App.Services = services.BuildServiceProvider();
         return base.CustomizeAppBuilder(builder).WithInterFont();
     }
 }